Cleanup: move logic for initializing Cytoscape event handlers to the hooks directory
